BAEL-1533 - Making a Spring MVC Form Remember Values
