+++ spring-5/pom.xml
@@ -39,13 +39,9 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-webflux</artifactId>
 		</dependency>
-		<dependency>
+        <dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-hateoas</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-thymeleaf</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.projectreactor</groupId>
@@ -56,6 +52,18 @@
 			<groupId>javax.json.bind</groupId>
 			<artifactId>javax.json.bind-api</artifactId>
 		</dependency>
+		<!-- Dependencies for Yasson -->
+		<!-- <dependency> -->
+		<!-- <groupId>org.eclipse</groupId> -->
+		<!-- <artifactId>yasson</artifactId> -->
+		<!-- <version>1.0</version> -->
+		<!-- </dependency> -->
+		<!-- <dependency> -->
+		<!-- <groupId>org.glassfish</groupId> -->
+		<!-- <artifactId>javax.json</artifactId> -->
+		<!-- <version>1.1.2</version> -->
+		<!-- </dependency> -->
+		<!-- Dependencies for Johnzon -->
 		<dependency>
 			<groupId>org.apache.geronimo.specs</groupId>
 			<artifactId>geronimo-json_1.1_spec</artifactId>
@@ -64,7 +72,6 @@
 		<dependency>
 			<groupId>org.apache.johnzon</groupId>
 			<artifactId>johnzon-jsonb</artifactId>
-			<version>${johnzon.version}</version>
 		</dependency>
 		<!-- utils -->
 		<dependency>
@@ -205,8 +212,8 @@
 		<java.version>1.8</java.version>
 		<junit.platform.version>1.0.0</junit.platform.version>
 		<maven-surefire-plugin.version>2.20</maven-surefire-plugin.version>
+		<spring.version>5.0.2.RELEASE</spring.version>
 		<reactor-spring.version>1.0.1.RELEASE</reactor-spring.version>
-		<johnzon.version>1.1.3</johnzon.version>
 		<geronimo-json_1.1_spec.version>1.0</geronimo-json_1.1_spec.version>
 		<asciidoctor-plugin.version>1.5.6</asciidoctor-plugin.version>
 		<snippetsDirectory>${project.build.directory}/generated-snippets</snippetsDirectory>
