--- spring-5/src/main/java/com/baeldung/sessionattrs/TodoList.java
@@ -1,8 +1,0 @@
-package com.baeldung.sessionattrs;
-
-import java.util.ArrayDeque;
-
-@SuppressWarnings("serial")
-public class TodoList extends ArrayDeque<TodoItem>{
-
-}
