--- spring-5/src/main/resources/templates/sessionattrs/sessionattributestodos.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity4" lang="en">
-    <head>
-        <title>Session Scope in Spring MVC</title>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-        <meta name="description" content="">
-        <meta name="author" content="">
-        <link href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/css/bootstrap.min.css" 
-            rel="stylesheet" 
-            integrity="sha384-/Y6pD6FV/Vv2HJnA6t+vslU6fwYXjCFtcEpHbNJ0lyAFsXTsjBbfaDjzALeQsN6M" 
-            crossorigin="anonymous">
-    </head>
-    <body>
-    	<div class="container">
-    		<p>
-    			<h3>Session Attributes Example</h3>
-    		</p>
-    	</div>
-        <div class="container">
-        	<br/>
-        	<div class="row">
-        				<a th:href="@{/sessionattributes/form}" class="btn btn-sm btn-primary btn-block" role="button" 
-										style="width: 10em;">Add New</a>
-        	</div>
-        	<br/>
-        	<div class="row">
-        		<div class="col card" style="padding: 1em;">
-       				<div class="card-block">
-       					<h5 class="card-title">TODO List</h5>
-						<table>
-							<tbody>
-								<tr>
-									<th width="75%">Description</th>
-									<th width="25%" >Create Date</th>
-								</tr>
-								<tr th:each="todo : ${todos}">
-									<td th:text="${todo.description}">Description</td>
-									<td th:text="${#temporals.format(todo.createDate, 'MM-dd-yyyy HH:mm:ss')}">Create Date</td>
-								</tr>
-							</tbody>
-						</table>
-					</div>
-				</div>
-			</div>
-        </div>
-    </body>
-</html>
