--- spring-5/src/test/java/com/baeldung/sessionattrs/TestConfig.java
@@ -1,17 +1,0 @@
-package com.baeldung.sessionattrs;
-
-import org.springframework.beans.factory.config.CustomScopeConfigurer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.SimpleThreadScope;
-
-@Configuration
-public class TestConfig {
-
-    @Bean
-    public CustomScopeConfigurer customScopeConfigurer() {
-        CustomScopeConfigurer configurer = new CustomScopeConfigurer();
-        configurer.addScope("session", new SimpleThreadScope());
-        return configurer;
-    }
-}
