+++ spring-mvc-forms-jsp/src/main/java/com/baeldung/springmvcforms/domain/Customer.java
