+++ spring-mvc-forms-jsp/src/main/webapp/WEB-INF/html/user.html
