+++ spring-mvc-forms-jsp/src/main/webapp/WEB-INF/views/customerView.jsp
