+++ spring-mvc-forms-thymeleaf/src/main/java/com/baeldung/sessionattrs/Config.java
@@ -1,0 +1,44 @@
+package com.baeldung.sessionattrs;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Scope;
+import org.springframework.context.annotation.ScopedProxyMode;
+import org.springframework.core.Ordered;
+import org.springframework.web.context.WebApplicationContext;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.thymeleaf.templatemode.TemplateMode;
+import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver;
+import org.thymeleaf.templateresolver.ITemplateResolver;
+
+@EnableWebMvc
+@Configuration
+public class Config implements WebMvcConfigurer {
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/").setViewName("index");
+        registry.setOrder(Ordered.HIGHEST_PRECEDENCE);
+    }
+
+    @Bean
+    @Scope(
+        value = WebApplicationContext.SCOPE_SESSION, 
+        proxyMode = ScopedProxyMode.TARGET_CLASS)
+    public TodoList todos() {
+        return new TodoList();
+    }
+
+    @Bean
+    public ITemplateResolver templateResolver() {
+        ClassLoaderTemplateResolver resolver 
+            = new ClassLoaderTemplateResolver();
+        resolver.setPrefix("templates/sessionattrs/");
+        resolver.setSuffix(".html");
+        resolver.setTemplateMode(TemplateMode.HTML);
+        resolver.setCharacterEncoding("UTF-8");
+        return resolver;
+    }
+}
