+++ spring-mvc-forms-thymeleaf/src/main/java/com/baeldung/sessionattrs/TodoList.java
@@ -1,0 +1,8 @@
+package com.baeldung.sessionattrs;
+
+import java.util.ArrayDeque;
+
+@SuppressWarnings("serial")
+public class TodoList extends ArrayDeque<TodoItem>{
+
+}
