+++ spring-mvc-forms-thymeleaf/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+server.port=8081
+
+logging.level.root=INFO
