+++ spring-mvc-forms-thymeleaf/src/main/resources/templates/sessionattrs/sessionattributesform.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity4" lang="en">
+    <head>
+        <title>Session Scope in Spring MVC</title>
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+        <meta name="description" content="">
+        <meta name="author" content="">
+        <link href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-/Y6pD6FV/Vv2HJnA6t+vslU6fwYXjCFtcEpHbNJ0lyAFsXTsjBbfaDjzALeQsN6M" crossorigin="anonymous">
+    </head>
+    <body>
+    	<div class="container">
+    		<p>
+    			<h3>Session Attributes Example</h3>
+    		</p>
+    	</div>
+        <div class="container">
+            <h5>Enter a TODO</h5>
+                <form action="#" th:action="@{/sessionattributes/form}" th:object="${todo}" method="post">
+                	<div class="input-group">
+                		<input type="text" th:field="*{description}" class="form-control" placeholder="TODO" required autofocus/>
+						<button class="btn btn-sm btn-primary form-control text-center" type="submit" style="max-width: 80px;">Create</button>
+                	</div>
+                </form>
+        </div>
+    </body>
+</html>
